Removed the 'Choose file' button, and added a (x) on Insulet uploads.
I believe this patch fixes https://trello.com/c/P0TpowaC
I've added some simple conditionals and styling to hide the 'choose file' button
once an ibf file has been selected in the uploader, as well as adding a 'delete'
icon button next to the file name.  That button removes the file from the list and
causes the choose and upload buttons to return to their default state by virtue of
toggling the 'isBlockModeFileChosen' state flag on the Upload component.

I have *not* tested that the proper file is uploaded to the server, but the current
upload behavior should be the same as that of clicking the 'choose file' button
multiple times.