Update message seen when usb cable bad
When I tried to upload my data, I had a cable that apparently only provides power and does not transfer signal.  After a bunch of troubles, I found the entry on http://support.tidepool.org/article/13-upload-your-dexcom-g4 about trying another USB cable and that fixed the problem.  Having that as an option in the error message would've likely been a nicer experience.