Only show "Four Weeks" menu option if debug enabled
Since the "Four Weeks" upload period is only possible for 600-series,
we don't want to make it generally available.
But it's also handy for development and QA purposes, so we still would
like to be able to access it somehow.
Now make it available only if the Debug Mode is enabled.

* Also fix linting warning