Fixes for 600-series calibrations
The code was previously using flags in SMBG readings to determine if the
event is a calibration. However, these flags only indicated an _intent_
to calibrate, and not whether the calibration was successful or not,
leading to discrepancies with CareLink reports.
Now using the CALIBRATION_COMPLETE event to register `deviceEvent`s of
subType `calibration`.