Comms error handling bug fix
Sets retries back to zero after issuing a "ResendPacketsCommand"
so that we don't spam the output command queue on the USB stick
with requests.
Doing so causes the stick to get buffer issues, and also means that we
double up on getting packets that we already had requested.

Also throw an InvalidStateError if we get packets we already received,
to make it really obvious that something is still not quite right.