Align wizard carb circles to bolus event
Carb circles for Normal Bolus events were still not quite aligned, since
a large bolus (say 14U), could still take over 10 minutes to deliver.
Now use the Bolus Programmed event to determine when the actual
bolus start is for all Bolus Delivered event types.
* Fixes https://trello.com/c/hZSvQjMZ

Also added a change so that a Basal Complete event without a matching
Basal Start event is not an error. This can happen either when the pump
history memory is full, and the most recent basal event is a basal end,
or if you're only requesting a delta (or 4 weeks).
Now just skip the basal event altogether.