Remove CCITT checksum from ComD payload
Having the CCITT checksum as part of the ComD payload caused some issues
for messages that were expecting to be able to read the whole payload,
and found an unwanted extra 2 bytes there (the checksum).
Also renamed the member variable to be consistent with its purpose.

* Fixes https://trello.com/c/S11Adtjz

This also includes a change to only leave the 802.15.4 network if we
have an ad hoc connection. Doing so all the time was causing linked
meters and pumps to become unlinked, which was bad.