Don't create temp basals when missing suppressed basal info
If the first basal event in the pump history is a temp basal, then there
will be no corresponding basal schedule by which we can model a temp
basal's suppressed info.
The driver should not create any temp basal events for which there are
no corresponding suppressed basal schedule events.

* Fixes https://trello.com/c/7hNQG2qR